Disable auto Stockfish downloads
Stockfish removed the downloads we were depending on and redirects to the new downloads page. This causes the engine resolution to fail with a cryptic error. So now we just throw early with a useful message to notify the user.

The new downloads are on GitHub actions and follow a different pattern with different available architectures. This makes updating the resolver non-trivial, hence I've just made it throw for now.